feat :: jooq manual context (r2dbc doesn't have autoconfig)
