--- src/main/kotlin/com/teamapi/palette/config/RedisConfig.kt
@@ -1,19 +1,0 @@
-package com.teamapi.palette.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.redis.connection.RedisConnectionFactory
-import org.springframework.data.redis.core.RedisTemplate
-import org.springframework.data.redis.repository.configuration.EnableRedisRepositories
-
-
-@Configuration
-@EnableRedisRepositories
-class RedisConfig {
-    @Bean
-    fun redisTemplate(redisConnectionFactory: RedisConnectionFactory): RedisTemplate<*, *> {
-        val template = RedisTemplate<String, String>()
-        template.connectionFactory = redisConnectionFactory
-        return template
-    }
-}
